--- e2e/test/scenarios/filters/reproductions/25927-column-filters-not-working-after-cc.cy.spec.js
@@ -1,53 +1,0 @@
-import { SAMPLE_DB_ID } from "e2e/support/cypress_data";
-import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-import {
-  popover,
-  restore,
-  tableHeaderClick,
-  visitQuestionAdhoc,
-} from "e2e/support/helpers";
-
-const { ORDERS, ORDERS_ID } = SAMPLE_DATABASE;
-
-const query = {
-  dataset_query: {
-    database: SAMPLE_DB_ID,
-    query: {
-      "source-query": {
-        "source-table": ORDERS_ID,
-        aggregation: [["count"]],
-        breakout: [["field", ORDERS.CREATED_AT, { "temporal-unit": "month" }]],
-      },
-      expressions: {
-        "Custom Count": ["field", "count", { "base-type": "type/Integer" }],
-      },
-    },
-    type: "query",
-  },
-  display: "table",
-};
-
-describe("issue 25927", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-    visitQuestionAdhoc(query);
-  });
-
-  it("column filter should work for questions with custom column (metabase#25927)", () => {
-    tableHeaderClick("Created At: Month");
-    popover().within(() => {
-      cy.findByText("Filter by this column").click();
-      cy.findByText("Last 30 days").click();
-    });
-
-    cy.wait("@dataset");
-
-    // Click on the filter again to try updating it
-    cy.findByTestId("qb-filters-panel")
-      .contains("Created At: Month is in the previous 30 days")
-      .click();
-
-    popover().button("Update filter").should("not.be.disabled");
-  });
-});
